// webapp/src/app/game/enums.ts
export enum ProductionBuilding {
  Unknown = 'Unknown',
  AdvancedCoffeeRoaster = 'Advanced Coffee Roaster',
  AdvancedCottonMill = 'Advanced Cotton Mill',
  AdvancedRumDistillery = 'Advanced Rum Distillery',
  AlpacaFarm = 'Alpaca Farm',
  Apiary = 'Apiary',
  ArcticGasMine = 'Arctic Gas Mine',
  ArtisanalKitchen = 'Artisanal Kitchen',
  AssemblyLineBiscuits = 'Assembly Line: Biscuits',
  AssemblyLineElevators = 'Assembly Line: Elevators',
  AssemblyLineTypewriters = 'Assembly Line: Typewriters',
  Bakery = 'Bakery',
  BearHuntingCabin = 'Bear Hunting Cabin',
  BicycleFactory = 'Bicycle Factory',
  BombinWeaver = 'Bombin Weaver',
  BrassSmeltery = 'Brass Smetlery',
  Brewery = 'Brewery',
  BrickDryHouse = 'Brick Dry-House',
  BrickFactory = 'Brick Factory',
  CabAssemblyLine = 'Cab Assembly Line',
  Cannery = 'Cannery',
  CaoutchoucPlantation = 'Caoutchouc Plantation',
  CaribouHuntingCabin = 'Caribou Hunting Cabin',
  CattleFarm = 'Cattle Farm',
  ChampagneCellar = 'Champagne Cellar',
  CharcoalKiln = 'Charcoal Kiln',
  ChemicalPlantEthanol = 'Chemical Plant: Ethanol',
  ChemicalPlantLacquer = 'Chemical Plant: Lacquer',
  ChemicalPlantLemonade = 'Chemical Plant: Lemonade',
  ChemicalPlantShampoo = 'Chemical Plant: Shampoo',
  ChemicalPlantSouvenirs = 'Chemical Plant: Souvenirs',
  Coachmakers = 'Coachmakers',
  ChocolateFactory = 'Chocolate Factory',
  CigarFactory = 'Cigar Factory',
  ClayPit = 'Clay Pit',
  Clockmakers = 'Clockmakers',
  CoalMine = 'Coal Mine',
  CocoaPlantation = 'Cocoa Plantation',
  CoffeePlantation = 'Coffee Plantation',
  CoffeeRoaster = 'Coffee Roaster',
  ConcreteFactory = 'Concrete Factory',
  CopperMine = 'Copper Mine',
  CornFarm = 'Corn Farm',
  CottonMill = 'Cotton Mill',
  CottonPlantation = 'Cotton Plantation',
  DynamiteFactory = 'Dynamite Factory',
  FeltProducer = 'Felt Producer',
  FilamentFactory = 'Filament Factory',
  Fishery = 'Fishery',
  FishOilFactory = 'Fish Oil Factory',
  FlourMill = 'Flour Mill',
  FrameworkKnitters = 'Framework Knitters',
  FriedPlantainKitchen = 'Fried Plantain Kitchen',
  FurDealer = 'Fur Dealer',
  Furnace = 'Furnace',
  Glassmakers = 'Glassmakers',
  GoatFarm = 'Goat Farm',
  Goldsmiths = 'Goldsmiths',
  GoldOreMine = 'Gold Ore Mine',
  GooseFarm = 'Goose Farm',
  GrainFarm = 'Grain Farm',
  GramophoneFactory = 'Gramophone Factory',
  HeavyWeaponsFactory = 'Heavy Weapons Factory',
  HibiscusFarm = 'Hibiscus Farm',
  HopFarm = 'Hop Farm',
  HuntingCabin = 'Hunting Cabin',
  HuskyFarm = 'Husky Farm',
  HuskySledFactory = 'Husky Sled Factory',
  IndigoFarm = 'Indigo Farm',
  IronMine = 'Iron Mine',
  Jewellers = 'Jewellers',
  LightBulbFactory = 'Light Bulb Factory',
  LimestoneQuarry = 'Limestone Mine',
  LinenMill = 'Linen Mill',
  LinseedFarm = 'Linseed Farm',
  LobsterFishery = 'Lobster Fishery',
  LumberjacksHut = 'Lumberjack\'s Hut',
  Malthouse = 'Malthouse',
  MarquetryWorkshop = 'Marquetry Workshop',
  MotorAssemblyLine = 'Motor Assembly Line',
  OilLampFactory = 'Oil Lamp Factory',
  OilRefinery = 'Oil Refinery',
  OrchardCherryWood = 'Orchard: Cherry Wood',
  OrchardCinnamon = 'Orchard: Cinnamon',
  OrchardCitrus = 'Orchard: Citrus',
  OrchardCoconutOil = 'Orchard: Coconut Oil',
  OrchardJam = 'Orchard: Jam',
  OrchardResin = 'Orchard: Resin',
  PaperMill = 'Paper Mill',
  ParkaFactory = 'Parka Factory',
  PearlFarm = 'Pearl Farm',
  PemmicanCookhouse = 'Pemmican Cookhouse',
  PigFarm = 'Pig Farm',
  PlantainPlantation = 'Plantain Plantation',
  PonchoDarner = 'Poncho Darner',
  PotatoFarm = 'Potato Farm',
  PrimeHuntingCabin = 'Prime Hunting Cabin',
  RedPepperFarm = 'Red Pepper Farm',
  RenderingWorks = 'Rendering Works',
  RumDistillery = 'Rum Distillery',
  Sailmakers = 'Sailmakers',
  SaltpeterWorks = 'Saltpeter Works',
  SaltWorks = 'SaltWorks',
  SandMine = 'Sand Mine',
  Sawmill = 'Sawmill',
  SchnappsDistillery = 'Schnapps Distillery',
  SealHuntingDocks = 'Seal Hunting Docks',
  SewingMachineFactory = 'Sewing Machine Factory',
  SheepFarm = 'Sheep Farm',
  Slaughterhouse = 'Slaughterhouse',
  SledFrameFactory = 'Sled Frame Factory',
  SleepingBagFactory = 'Sleeping Bag Factory',
  SoapFactory = 'Soap Factory',
  SpectacleFactory = 'Spectacle Factory',
  SpiceFarm = 'Spice Farm',
  Steelworks = 'Steelworks',
  SugarCanePlantation = 'Sugar Cane Plantation',
  SugarRefinery = 'Sugar Refinery',
  TailorsShop = 'Tailor\'s Shop',
  TapestryLooms = 'Tapestry Looms',
  TeaSpicer = 'Tea Spicer',
  TeffFarm = 'Teff Farm',
  TelephoneManufacturer = 'Telephone Manufacturer',
  TobaccoPlantation = 'Tobacco Plantation',
  TortillaMaker = 'Tortilla Maker',
  Vineyard = 'Vineyard',
  WeaponFactory = 'Weapon Factory',
  WhalingStation = 'Whaling Station',
  WindowMakers = 'Window Makers',
  ZincMine = 'Zinc Mine',
};

export enum ProductionType {
  Mine = 'Mine',
  Factory = 'Factory',
  PlantFarm = 'Plant Farm',
  AnimalFarm = 'Animal Farm',
  HuntingCabin = 'Hunting Cabin',
  Orchard = 'Orchard',
  Oil = 'Oil',
};

export enum Good {
  Unknown = 'Unknown',
  AdvancedWeapons = 'Advanced Weapons',
  AlpacaWool = 'Alpaca Wool',
  AluminiumProfiles = 'Aluminium Profiles',
  ArcticGas = 'Arctic Gas',
  Bauxite = 'Bauxite',
  BearFur = 'Bear Fur',
  Beef = 'Beef',
  Beer = 'Beer',
  Beeswax = 'Beeswax',
  Biscuits = 'Biscuits',
  Bombs = 'Bombs',
  BowlerHats = 'Bowler Hats',
  Brass = 'Brass',
  Bread = 'Bread',
  Bricks = 'Bricks',
  CamphorWax = 'Camphor Wax',
  CannedFood = 'Canned Food',
  Caoutchouc = 'Caoutchouc',
  CaribouMeat = 'Caribou Meat',
  Cement = 'Cement',
  Champagne = 'Champagne',
  CherryWood = 'Cherry Wood',
  Chocolate = 'Chocolate',
  Cigars = 'Cigars',
  Cinnamon = 'Cinnamon',
  Citrus = 'Citrus',
  Clay = 'Clay',
  Chassis = 'Chassis',
  Coal = 'Coal',
  Cocoa = 'Cocoa',
  CoconutOil = 'Coconut Oil',
  Coffee = 'Coffee',
  CoffeeBeans = 'Coffee Beans',
  Copper = 'Copper',
  Corn = 'Corn',
  Cotton = 'Cotton',
  CottonFabric = 'Cotton Fabric',
  Dynamite = 'Dynamite',
  Elevators = 'Elevators',
  Ethanol = 'Ethanol',
  Felt = 'Felt',
  Filaments = 'Filaments',
  Fish = 'Fish',
  FishOil = 'Fish Oil',
  Flour = 'Flour',
  FriedPlantains = 'Fried Plantains',
  FurCoats = 'Fur Coats',
  Furs = 'Furs',
  Glass = 'Glass',
  Glasses = 'Glasses',
  GoatMilk = 'Goat Milk',
  Gold = 'Gold',
  GoldOre = 'Gold Ore',
  GooseFeathers = 'Goose Feathers',
  Goulash = 'Goulash',
  Grain = 'Grain',
  Grapes = 'Grapes',
  Gramophones = 'Gramophones',
  Helium = 'Helium',
  Herbs = 'Herbs',
  HibiscusPetals = 'Hibiscus Petals',
  HibiscusTea = 'Hibiscus Tea',
  Hops = 'Hops',
  Huskies = 'Huskies',
  HuskySleds = 'Husky Sleds',
  IceCream = 'Ice Cream',
  IndigoDye = 'Indigo Dye',
  Iron = 'Iron',
  Jam = 'Jam',
  Jewellery = 'Jewellery',
  Lacquer = 'Lacquer',
  LeatherBoots = 'Leather Boots',
  Lemonade = 'Lemonade',
  LightBulbs = 'Light Bulbs',
  Linen = 'Linen',
  Linseed = 'Linseed',
  Lobster = 'Lobster',
  Malt = 'Malt',
  Milk = 'Milk',
  Minerals = 'Minerals',
  MudBricks = 'Mud Bricks',
  NanduFeathers = 'Nandu Feathers',
  NanduLeather = 'Nandu Leather',
  Oil = 'Oil',
  OilLamps = 'Oil Lamps',
  Pamphlets = 'Pamphlets',
  Paper = 'Paper',
  Parkas = 'Parkas',
  Pearls = 'Pearls',
  Pemmican = 'Pemmican',
  PennyFarthings = 'Penny Farthings',
  Perfumes = 'Perfumes',
  Pigs = 'Pigs',
  Plantains = 'Plantains',
  PocketWatches = 'Pocket Watches',
  Ponchos = 'Ponchos',
  Potatoes = 'Potatoes',
  QuartzSand = 'Quartz Sand',
  RedPeppers = 'Red Peppers',
  ReinforcedConcrete = 'Reinforced Concrete',
  Resin = 'Resin',
  Rum = 'Rum',
  Sails = 'Sails',
  Salt = 'Salt',
  Saltpetre = 'Saltpetre',
  Sausages = 'Sausages',
  Schnapps = 'Schnapps',
  SealSkin = 'Seal Skin',
  SewingMachines = 'Sewing Machines',
  Shampoo = 'Shampoo',
  Sleds = 'Sleds',
  SleepingBags = 'Sleeping Bags',
  Soap = 'Soap',
  Souvenirs = 'Souvenirs',
  Spices = 'Spices',
  SteamCarriages = 'Steam Carriages',
  SteamMotors = 'Steam Motors',
  Steel = 'Steel',
  SteelBeams = 'Steel Beams',
  Sugar = 'Sugar',
  SugarCane = 'Sugar Cane',
  TailoredSuits = 'Tailored Suits',
  Tallow = 'Tallow',
  Tapestries = 'Tapestries',
  Teff = 'Teff',
  Telephones = 'Telephones',
  Timber = 'Timber',
  Tobacco = 'Tobacco',
  Tortillas = 'Tortillas',
  Toys = 'Toys',
  Typewriters = 'Typewriters',
  Violins = 'Violins',
  WaterDrop = 'Water Drop',
  Weapons = 'Weapons',
  WhaleOil = 'Whale Oil',
  Windows = 'Windows',
  Wood = 'Wood',
  WoodVeneers = 'Wood Veneers',
  Wool = 'Wool',
  WorkClothes = 'Work Clothes',
  Zinc = 'Zinc',
};

export enum Boost {
  None = 'None',
  Electricity = 'Electricity',
  TracktorBarn = 'Tracktor Barn',
  Fertilizer = 'Fertilizer',
  Silo = 'Silo',
};

export enum DepartmentOfLaborPolicy {
  None = 'None',
  LandReformAct = 'Land Reform Act',
  SkilledLaborAct = 'Skilled Labor Act',
  GalvanicGrantsAct = 'Galvanic Grants Act',
  FactoryInspectionsAct = 'Factory Inspections Act',
  UnionSubsidiesAct = 'Union Subsidies Act',
};

export enum Region {
  Unknown = 'Unknown',
  OldWorld = 'Old World',
  NewWorld = 'New World',
  CapeTrelawney = 'Cape Trelawney',
  Arctic = 'The Arctic',
  Enbesa = 'Enbesa',
};